Replace TrackerService with LocationService
Improve reliability of background location tracking
Pass handling of location updates to LocationReceiver
Start different service in MainFragment
Add details to AndroidManifest.xml
