+++ app/src/main/java/com/chornsby/touristtracker/MainFragment.java
@@ -83,7 +83,7 @@ public class MainFragment extends Fragment {
         });
 
         // Start Service without defined action in order to respect SharedPreferences
-        Intent intent = new Intent(getActivity(), TrackerService.class);
+        Intent intent = new Intent(getActivity(), LocationService.class);
         getActivity().startService(intent);
 
         return rootView;
